feat(LottoResult): 수익률 소수점 둘째자리 까지 출력되도록 수정 (테스트코드 성공

+++ src/main/java/step2/domain/LottoResult.java
@@ -24,6 +24,6 @@ public class LottoResult {
     }
 
     public double getYield(Money money) {
-        return getTotalPrize() / Long.valueOf(money.getMoney());
+        return getTotalPrize() / Double.valueOf(money.getMoney());
     }
 }
